backend/requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.3
conan==2.0.1
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
django_vite_plugin==1.3.0
djangorestframework==3.14.0
executing==1.2.0
fasteners==0.18
fastjsonschema==2.16.3
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
graphviz==0.20.1
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
jaraco.context==4.3.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
neat-python==0.92
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
patch-ng==1.17.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.7.23
platformdirs==3.9.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.1.3
Pygments==2.15.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.24.2
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
whitenoise==6.5.0
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
